% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBAT.list_versions.R
\name{GBAT.list_versions}
\alias{GBAT.list_versions}
\title{Available versions of DCP's Geosupport.}
\usage{
GBAT.list_versions()
}
\value{
A data frame containing available versions of DCP's Geosupport geocoding software.
}
\description{
The \code{GBAT.list_versions} provides a data frame containing available versions of DCP's Geosupport geocoding software which were downloaded from \href{https://www1.nyc.gov/site/planning/data-maps/open-data.page#geocoding_application}{DCP BYTES of the BIG APPLE}.
}
\details{
Geosupport can be downloaded from \href{https://www1.nyc.gov/site/planning/data-maps/open-data.page#geocoding_application}{DCP BYTES of the BIG APPLE}. Excellent documentation is provide in the \href{https://nycplanning.github.io/Geosupport-UPG/}{Geosupport System User Programming Guide}.
}
\examples{

#available GBAT versions installed on R server
gc_versions <- GBAT.list_versions()
}
